Move paths to one file
