+++ app/manage/lib/core/router/manage_route_path.dart
@@ -1,1 +1,11 @@
 abstract class ManageRoutePath {}
+
+class ManageTeamPath extends ManageRoutePath {
+  final String id;
+
+  ManageTeamPath(this.id);
+}
+
+class ManageTeamsPath extends ManageRoutePath {}
+
+class ManageUnknownPath extends ManageRoutePath {}
