--- app/manage/lib/core/router/manage_team_path.dart
@@ -1,7 +1,0 @@
-import 'package:manage/core/router/manage_route_path.dart';
-
-class ManageTeamPath extends ManageRoutePath {
-  final String id;
-
-  ManageTeamPath(this.id);
-}
