--- app/manage/lib/core/router/manage_teams_path.dart
@@ -1,3 +1,0 @@
-import 'package:manage/core/router/manage_route_path.dart';
-
-class ManageTeamsPath extends ManageRoutePath {}
